refactor: 타입 types 폴더로 분리
